refactor: lift Prod-side conversion impls onto per-typegen emitConvert (Phase 5)
Move all three Prod conversion bodies out of ProdDispatch:

  - prodCaseImpl              → ProductCaseSirTypeGenerator.emitConvert (~440 LOC)
  - prodCaseUplcConstrImpl    → ProductCaseUplcConstrSirTypeGenerator.emitConvert
  - prodCaseOneElementImpl    → OneElementWrapperEmitter.emitConvert
                                (gen → this; case-class instance method)

Cross-class delegations resolve via the public emitConvert entry points:
ProductCaseUplcConstrSirTypeGenerator and OneElementWrapperEmitter both
fall through to ProductCaseSirTypeGenerator.emitConvert for the
non-typegen-specific arms. The OneElementWrapper-source arm inside
ProductCaseSirTypeGenerator.emitConvert calls the wrapper instance's
own emitConvert (no longer the dispatcher private fn).

ProdDispatch.toRepresentation routes each typegen to its emitConvert.
Net ProdDispatch shrink: ~570 lines; unused imports (Doc / Term /
UplcAnnotation / MajorProtocolVersion / Builder) cleaned up.
